refactor(base): reorganize package contents
* co-locate errors
* use fewer files
* remove unnecessary ValidateDataset function
