+++ base/base.go
@@ -1,33 +1,8 @@
-// Package base defines business that operates on local data.
-// it's main job is to composing APIs from the lower half of our tech stack,
-// providing uniform functions for higher up packages, mainly p2p and lib.
-// p2p and lib use base as the only way of operate on the local repo
-// Here's some ascii art to clarify the stack:
-//
-//    ┌───────────────┐ ┌───────────────┐
-//    │      cmd      │ │     api       │
-//    └───────────────┘ └───────────────┘
-//    ┌─────────────────────────────────┐
-//    │               lib               │
-//    └─────────────────────────────────┘
-//    ┌───────────────────────┐
-//    │          p2p          │
-//    └───────────────────────┘
-//    ┌─────────────────────────────────┐
-//    │              base               │  <-- you are here
-//    └─────────────────────────────────┘
-//    ┌──────┐ ┌──────┐ ┌──────┐ ┌──────┐
-//    │ repo │ │ dsfs │ │ qfs  │ │ ...  │
-//    └──────┘ └──────┘ └──────┘ └──────┘
-//
-// There are packages omitted from this diagram, but these are the vitals.
-// base functions mainly work with repo.Repo instances, using repo interface methods
-// and related packages to do their work. This is part of a larger pattern of
-// having lib rely on lower level interfaces wherever possible to enhance
-// configurability
+// Package base defines functions that operate on local data
 package base
 
 import (
+	"fmt"
 	"time"
 
 	golog "github.com/ipfs/go-log"
@@ -42,4 +17,14 @@ var (
 	// files unresolved.
 	// TODO (b5) - allow -1 duration as a sentinel value for no timeout
 	OpenFileTimeoutDuration = time.Millisecond * 250
+)
+
+var (
+	// ErrUnlistableReferences is an error for when listing references encounters
+	// some problem, but these problems are non-fatal
+	ErrUnlistableReferences = fmt.Errorf("Warning: Some datasets could not be listed, because of invalid state. These datasets still exist in your repository, but have references that cannot be resolved. This will be fixed in a future version. You can see all datasets by using `qri list --raw`")
+	// ErrNameTaken is an error for when a name for a new dataset is already being used
+	ErrNameTaken = fmt.Errorf("name already in use")
+	// ErrDatasetLogTimeout is an error for when getting the datasetLog times out
+	ErrDatasetLogTimeout = fmt.Errorf("datasetLog: timeout")
 )
